restructured src dir
split in components and pages dir

- components handle raw pieces
- pages hold respective website pages
- also fixed navbar logo link
