+++ src/components/Navbar/Navbar.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import logo from "../logo.png";
+import logo from "../../logo.png";
 import {
   Collapse,
   Navbar,
